+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-* 2.0.0 (...)
+* 2.0.0 (2013-)
 
-  * Migrated to Solarium 3.0, use the 1.x version of this bundle if you want to keep using Solarium 2
+  * Migrated to Solarium 3.x, use the 1.x version of this bundle if you want to keep using Solarium 2.x
+  * Added support for defining multiple endpoints
+  * Added Monolog support
+  * Added Stopwatch data for the Symfony Profiler
 
 * 1.1.0 (2013-01-07)